[map_filter] Fix codeclimate issue
